+++ src/image/mix/with_colormap.rs
@@ -1,5 +1,3 @@
-pub mod with_colormap;
-
 use embedded_graphics::Drawable;
 use embedded_graphics::draw_target::DrawTarget;
 use embedded_graphics::geometry::{Dimensions, OriginDimensions, Size};
@@ -8,15 +6,40 @@ use embedded_graphics::pixelcolor::{BinaryColor, Gray2, Gray4, Gray8, PixelColor
 use embedded_graphics::primitives::Rectangle;
 
 use crate::color::{Colormap, Invert, Screen, WeightedAvg};
-use crate::image::{Colors, Image, SubImage};
+use crate::image::{Colors, Image, Mixed, SubImage};
 
-/// Image with references to two overlapping image drawables and a colormap.
+/// Association with a colormap.
+pub trait WithColormap<'a, T: Copy, const N: usize> {
+    /// The output type.
+    type Output<'b>
+    where
+        T: 'b,
+        Self: 'a;
+
+    /// Returns the association with a colormap, which then belongs to the image.
+    fn with_colormap<'b>(&'a self, colormap: &'b Colormap<T, N>) -> Self::Output<'b>;
+}
+
+/// Association between an image and a colormap, which implements [`Mixed`] and allows for the
+/// images being color-mixed to each use their own colormaps.
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct WithColormapImage<'a, 'b, U, T, const N: usize>
+where
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
+    U: ImageDrawable + Colors<U::Color>,
+{
+    image: &'a Image<U>,
+    colormap: &'b Colormap<T, N>,
+}
+
+/// Image with references to two overlapping image drawables and a set of two colormaps.
 ///
 /// While also performing color conversion, drawing this image drawable involves mixing the colors
-/// that form pairs of pixels in [`Screen`] blend mode.
+/// that form pairs of pixels in [`WeightedAvg`] blend mode.
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct ImageMix<'a, 'b, 'c, U, V, T, const N: usize>
+pub struct WithColormapImageMix<'a, 'b, 'c, 'd, U, V, T, const N: usize>
 where
     T: PixelColor + Default + Invert + Screen + WeightedAvg,
     U: ImageDrawable + Colors<U::Color>,
@@ -25,57 +48,55 @@ where
     first: Image<SubImage<'a, U>>,
     second: Image<SubImage<'b, V>>,
     colormap: &'c Colormap<T, N>,
+    other_colormap: &'d Colormap<T, N>,
     area: Rectangle,
 }
 
-/// Image drawable color-mixed with another image drawable in the area that is their intersection.
-pub trait Mixed<'a, U, T, C, const N: usize>
+impl<'a, U, T: Copy, const N: usize> WithColormap<'a, T, N> for Image<U>
 where
-    C: PixelColor + From<C::Raw>,
     T: PixelColor + Default + Invert + Screen + WeightedAvg,
-    U: ImageDrawable<Color = C> + Colors<C>,
+    U: ImageDrawable + Colors<U::Color>,
 {
-    /// The output type.
-    type Output<'b, 'c, V>
+    type Output<'b>
+        = WithColormapImage<'a, 'b, U, T, N>
     where
-        T: 'c,
-        V: ImageDrawable<Color = C> + Colors<C> + 'b,
+        T: 'b,
         Self: 'a;
 
-    /// Returns the image drawable's intersection with the specified image, color-mixed in the
-    /// resulting area.
-    fn mixed<'b, 'c, V>(
-        &'a self,
-        other: &'b Image<V>,
-        colormap: &'c Colormap<T, N>,
-    ) -> Self::Output<'b, 'c, V>
-    where
-        V: ImageDrawable<Color = C> + Colors<C>;
+    fn with_colormap<'b>(&'a self, colormap: &'b Colormap<T, N>) -> Self::Output<'b> {
+        Self::Output {
+            image: self,
+            colormap,
+        }
+    }
 }
 
-impl<'a, 'b, 'c, U, V, T, const N: usize> ImageMix<'a, 'b, 'c, U, V, T, N>
+impl<'a, 'b, 'c, 'd, U, V, T, const N: usize> WithColormapImageMix<'a, 'b, 'c, 'd, U, V, T, N>
 where
     T: PixelColor + Default + Invert + Screen + WeightedAvg,
     U: ImageDrawable + Colors<U::Color>,
     V: ImageDrawable + Colors<V::Color>,
 {
-    /// Creates a new image drawable with two pre-cut image drawables and the specified colormap.
+    /// Creates a new image drawable with two pre-cut image drawables and the specified set of two
+    /// colormaps.
     const fn new(
         first: Image<SubImage<'a, U>>,
         second: Image<SubImage<'b, V>>,
         colormap: &'c Colormap<T, N>,
+        other_colormap: &'d Colormap<T, N>,
         area: Rectangle,
     ) -> Self {
         Self {
             first,
             second,
             colormap,
+            other_colormap,
             area,
         }
     }
 }
 
-impl<U, V, T, const N: usize> OriginDimensions for ImageMix<'_, '_, '_, U, V, T, N>
+impl<U, V, T, const N: usize> OriginDimensions for WithColormapImageMix<'_, '_, '_, '_, U, V, T, N>
 where
     T: PixelColor + Default + Invert + Screen + WeightedAvg,
     U: ImageDrawable + Colors<U::Color>,
@@ -93,7 +114,7 @@ macro_rules! impl_drawable {
         )*
     ) => {
         $(
-            impl<U, V, T> Drawable for ImageMix<'_, '_, '_, U, V, T, $array_length>
+            impl<U, V, T> Drawable for WithColormapImageMix<'_, '_, '_, '_, U, V, T, $array_length>
             where
                 T: PixelColor + Default + Invert + Screen + WeightedAvg,
                 U: ImageDrawable<Color = $color_type> + Colors<$color_type>,
@@ -116,19 +137,24 @@ macro_rules! impl_drawable {
                         .second
                         .colors()
                         .into_iter()
-                        .map(|color| self.colormap.get(color));
+                        .map(|color| self.other_colormap.get(color));
 
                     let start = self.colormap.first();
                     let end = self.colormap.last();
-                    if start == end {
-                        target.fill_solid(&self.area, start)
-                    } else {
-                        let colors = first
-                            .zip(second)
-                            .map(|(first, second)| first.screen(second, start, end));
+                    let other_start = self.other_colormap.first();
+                    let other_end = self.other_colormap.last();
+                    let colors = first
+                        .zip(second)
+                        .map(|(first, second)| {
+                            first.weighted_avg(
+                                second,
+                                start,
+                                end,
+                                other_start,
+                                other_end)
+                        });
 
-                        target.fill_contiguous(&self.area, colors)
-                    }
+                    target.fill_contiguous(&self.area, colors)
                 }
             }
         )*
@@ -149,31 +175,33 @@ macro_rules! impl_mixed {
         )*
     ) => {
         $(
-            impl<'a, U, T> Mixed<'a, U, T, $color_type, $array_length> for Image<U>
+            impl<'a, 'c, U, T> Mixed<'a, U, T, $color_type, $array_length>
+                for WithColormapImage<'a, 'c, U, T, $array_length>
             where
                 T: PixelColor + Default + Invert + Screen + WeightedAvg,
                 U: ImageDrawable<Color = $color_type> + Colors<$color_type>,
             {
-                type Output<'b, 'c, V>
-                    = ImageMix<'a, 'b, 'c, U, V, T, $array_length>
+                type Output<'b, 'd, V>
+                    = WithColormapImageMix<'a, 'b, 'c, 'd, U, V, T, $array_length>
                 where
-                    T: 'c,
+                    T: 'd,
                     V: ImageDrawable<Color = $color_type> + Colors<$color_type> + 'b,
                     Self: 'a;
 
-                fn mixed<'b, 'c, V>(
+                fn mixed<'b, 'd, V>(
                     &'a self,
                     other: &'b Image<V>,
-                    colormap: &'c Colormap<T, $array_length>,
-                ) -> Self::Output<'b, 'c, V>
+                    other_colormap: &'d Colormap<T, $array_length>,
+                ) -> Self::Output<'b, 'd, V>
                 where
                     V: ImageDrawable<Color = $color_type> + Colors<$color_type>,
                 {
-                    let area = self.bounding_box().intersection(&other.bounding_box());
-                    let first = self.clipped(&area);
+                    let area = self.image.bounding_box().intersection(&other.bounding_box());
+                    let first = self.image.clipped(&area);
                     let second = other.clipped(&area);
+                    let colormap = self.colormap;
 
-                    ImageMix::new(first, second, colormap, area)
+                    WithColormapImageMix::new(first, second, colormap, other_colormap, area)
                 }
             }
         )*