+++ src/style.rs
@@ -11,6 +11,7 @@
 //!   defaults to black; this color is filled in from top to bottom, for the length of the text run.
 //! </div>
 
+use core::cell::RefCell;
 use core::iter;
 
 use embedded_graphics::Drawable;
@@ -19,19 +20,22 @@ use embedded_graphics::geometry::{Dimensions, Point, Size};
 use embedded_graphics::iterator::raw::RawDataSlice;
 use embedded_graphics::pixelcolor::raw::BigEndian;
 use embedded_graphics::pixelcolor::{BinaryColor, Gray2, Gray4, Gray8, PixelColor};
-use embedded_graphics::primitives::{PrimitiveStyle, Rectangle, StyledDrawable};
+use embedded_graphics::primitives::{Primitive, PrimitiveStyle, Rectangle, Styled, StyledDrawable};
 use embedded_graphics::text::renderer::{CharacterStyle, TextMetrics, TextRenderer};
 use embedded_graphics::text::{Baseline, DecorationColor};
 
 use crate::adapter::DrawTargetExt;
 use crate::charmap::{Charmap, CharmapEntry};
-use crate::color::{Colormap, Invert, Linear, Screen};
+use crate::color::{Colormap, Invert, Linear, Screen, WeightedAvg};
 use crate::font::BitmapFont;
 use crate::glyph::NextGlyph;
-use crate::image::{Image, ImageRaw, Mixed};
+use crate::image::{Image, ImageRaw, Mixed, WithColormap};
+use crate::metrics::DecorationDimensions;
 use crate::rect::RectangleExt;
 
 pub use crate::builder::BitmapFontStyleBuilder;
+
+type StyledRectangle<T> = Styled<Rectangle, PrimitiveStyle<T>>;
 
 /// Style using a bitmap font.
 #[derive(Debug, Clone)]
@@ -39,7 +43,7 @@ pub use crate::builder::BitmapFontStyleBuilder;
 pub struct BitmapFontStyle<'a, 'b, T, C, const N: usize>
 where
     C: PixelColor + From<C::Raw>,
-    T: PixelColor + Default + Invert + Screen,
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
     RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
 {
     /// The bitmap font.
@@ -52,12 +56,41 @@ where
     pub underline_color: DecorationColor<T>,
     /// The strikethrough color.
     pub strikethrough_color: DecorationColor<T>,
+    /// The carryover from a previous call to either the [draw_string](Self::draw_string) method or
+    /// the [draw_whitespace](Self::draw_whitespace) method.
+    carryover: RefCell<Option<Carryover<'a, T, C, 2>>>,
+}
+
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+struct Carryover<'a, T, C, const N: usize>
+where
+    C: PixelColor + From<C::Raw>,
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
+    RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
+{
+    previous_image_colorable: Option<PreviousImageColorable<'a, T, C>>,
+    decorations: [Option<StyledRectangle<T>>; N],
+    line_piece: Rectangle,
+}
+
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+struct PreviousImageColorable<'a, T, C>
+where
+    C: PixelColor + From<C::Raw>,
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
+    RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
+{
+    previous_image: Image<ImageRaw<'a, C>>,
+    text_color: T,
+    background_color: T,
 }
 
 impl<'a, 'b, T, C, const N: usize> BitmapFontStyle<'a, 'b, T, C, N>
 where
     C: PixelColor + From<C::Raw>,
-    T: PixelColor + Default + Invert + Screen,
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
     RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
 {
     /// Creates a new style with the specified bitmap font and text color.
@@ -68,13 +101,26 @@ where
             .build()
     }
 
+    /// Creates a new style with the invisible bitmap font, the default text and background colors,
+    /// which are not specified, as are the decoration colors for text underline and strikethrough.
+    pub const fn const_default() -> Self {
+        Self {
+            font: &BitmapFont::NULL,
+            text_color: None,
+            background_color: None,
+            underline_color: DecorationColor::None,
+            strikethrough_color: DecorationColor::None,
+            carryover: RefCell::new(None),
+        }
+    }
+
     /// Returns the text color, falling back to the inverse of the default value for type `T` when
     /// not set to a value.
     fn text_color(&self) -> T {
         self.text_color.unwrap_or(T::default().invert())
     }
 
-    /// Returns the background_color, falling back to the default value for type `T` when not set
+    /// Returns the background color, falling back to the default value for type `T` when not set
     /// to a value.
     fn background_color(&self) -> T {
         self.background_color.unwrap_or_default()
@@ -99,12 +145,125 @@ where
             DecorationColor::Custom(color) => Some(color),
         }
     }
+
+    fn decorations_drawn<D>(
+        &self,
+        left: i32,
+        baseline: i32,
+        width: u32,
+        target: &mut D,
+    ) -> Result<[Option<StyledRectangle<T>>; 2], D::Error>
+    where
+        D: DrawTarget<Color = T>,
+    {
+        let stroke = |dimensions: DecorationDimensions, stroke_color: Option<_>| {
+            stroke_color.map(|stroke_color| {
+                let top = baseline.saturating_sub(dimensions.y_offset());
+                let height = dimensions.stroke_width();
+                let stroke_style = PrimitiveStyle::with_fill(stroke_color);
+                let stroke = Rectangle {
+                    top_left: Point::new(left, top),
+                    size: Size::new(width, height),
+                };
+
+                stroke.into_styled(stroke_style)
+            })
+        };
+
+        let decorations = [
+            stroke(self.font.underline, self.underline_color()),
+            stroke(self.font.strikethrough, self.strikethrough_color()),
+        ];
+
+        for decoration in decorations.into_iter().flatten() {
+            decoration.draw(target)?;
+        }
+
+        Ok(decorations)
+    }
+}
+
+macro_rules! impl_carryover {
+    (
+        $(
+            $color_type:ty, $array_length:literal,
+        )*
+    ) => {
+        $(
+            impl<'a, 'b, T, const N: usize> Carryover<'a, T, $color_type, N>
+            where
+                T: PixelColor + Default + Invert + Screen + WeightedAvg,
+                Colormap<T, $array_length>: Linear<T>,
+            {
+                fn redraw_whitespace<D, const M: usize>(
+                    &self,
+                    style: &BitmapFontStyle<'a, 'b, T, $color_type, M>,
+                    line_piece: Rectangle,
+                    target: &mut D,
+                ) -> Result<(), D::Error>
+                where
+                    D: DrawTarget<Color = T>,
+                {
+                    let background_color = style.background_color();
+                    let background_style = PrimitiveStyle::with_fill(background_color);
+                    let intersection = line_piece.intersection(&self.line_piece);
+                    if !intersection.is_zero_sized() {
+                        if let Some(previous_image_colorable) = &self.previous_image_colorable {
+                            let previous_image = &previous_image_colorable.previous_image;
+                            let previous_image_box = previous_image.bounding_box();
+                            if background_color != previous_image_colorable.background_color {
+                                let text_color = previous_image_colorable.text_color;
+                                let colormap = Colormap::linear(background_color, text_color);
+                                let mut adapter = target.value_mapped(&colormap);
+                                previous_image.clipped(&intersection).draw(&mut adapter)?;
+
+                                let above = intersection.above(&previous_image_box);
+                                let below = intersection.below(&previous_image_box);
+                                for fill_area in [above, below] {
+                                    fill_area.draw_styled(&background_style, target)?;
+                                }
+                            }
+                        } else {
+                            intersection.draw_styled(&background_style, target)?;
+                        }
+
+                        let decorations = self.decorations.into_iter().flatten();
+                        let decorations = decorations.map(|Styled { primitive, style }| {
+                            primitive.intersection(&intersection).into_styled(style)
+                        });
+
+                        for decoration in decorations {
+                            decoration.draw(target)?;
+                        }
+                    }
+
+                    let left = line_piece.left_of(&self.line_piece);
+                    let right = line_piece.right_of(&self.line_piece);
+                    let middle = line_piece.left_of(&right).right_of(&left);
+                    let above = middle.above(&self.line_piece);
+                    let below = middle.below(&self.line_piece);
+                    for fill_area in [left, right, above, below] {
+                        fill_area.draw_styled(&background_style, target)?;
+                    }
+
+                    Ok(())
+                }
+            }
+        )*
+    }
+}
+
+impl_carryover! {
+    BinaryColor, 2,
+    Gray2, 4,
+    Gray4, 16,
+    Gray8, 256,
 }
 
 impl<'a, T, C, const N: usize> CharacterStyle for BitmapFontStyle<'a, '_, T, C, N>
 where
     C: PixelColor + From<C::Raw>,
-    T: PixelColor + Default + Invert + Screen,
+    T: PixelColor + Default + Invert + Screen + WeightedAvg,
     RawDataSlice<'a, C::Raw, BigEndian>: IntoIterator<Item = C::Raw>,
 {
     type Color = T;
@@ -135,7 +294,7 @@ macro_rules! impl_text_renderer {
         $(
             impl<T, const N: usize> TextRenderer for BitmapFontStyle<'_, '_, T, $color_type, N>
             where
-                T: PixelColor + Default + Invert + Screen,
+                T: PixelColor + Default + Invert + Screen + WeightedAvg,
                 Colormap<T, $array_length>: Linear<T>,
             {
                 type Color = T;
@@ -226,6 +385,80 @@ macro_rules! impl_text_renderer {
                             image.mixed(previous_image, &colormap).draw(target)?;
 
                             image_box.right_of(&previous_image_box)
+                        } else if let Some(carryover) = self.carryover.take() {
+                            carryover.redraw_whitespace(self, line_piece, target)?;
+
+                            let image_box = image_box.left_half();
+                            let line_piece = image_box.y_extend(top, bottom);
+                            let intersection = line_piece.intersection(&carryover.line_piece);
+                            if !intersection.is_zero_sized() {
+                                let previous_image_colorable = carryover.previous_image_colorable;
+                                if let Some(previous_image_colorable) = previous_image_colorable {
+                                    let previous_image = &previous_image_colorable.previous_image;
+                                    let previous_image_box = previous_image.bounding_box();
+                                    let above = line_piece.above(&previous_image_box);
+                                    let below = line_piece.below(&previous_image_box);
+                                    for clip_area in [above, below] {
+                                        let mut adapter = target.value_mapped(&colormap);
+                                        image.clipped(&clip_area).draw(&mut adapter)?;
+                                    }
+
+                                    let image = image.with_colormap(&colormap);
+                                    let background_color = self.background_color();
+                                    let text_color = previous_image_colorable.text_color;
+                                    let colormap = Colormap::linear(background_color, text_color);
+                                    let above = intersection.above(&image_box);
+                                    let below = intersection.below(&image_box);
+                                    for clip_area in [above, below] {
+                                        let mut adapter = target.value_mapped(&colormap);
+                                        previous_image.clipped(&clip_area).draw(&mut adapter)?;
+
+                                        let above = clip_area.above(&previous_image_box);
+                                        let below = clip_area.below(&previous_image_box);
+                                        for fill_area in [above, below] {
+                                            fill_area.draw_styled(&background_style, target)?;
+                                        }
+                                    }
+
+                                    image.mixed(previous_image, &colormap).draw(target)?;
+                                } else {
+                                    let mut adapter = target.value_mapped(&colormap);
+                                    image.clipped(&intersection).draw(&mut adapter)?;
+
+                                    let above = intersection.above(&image_box);
+                                    let below = intersection.below(&image_box);
+                                    for fill_area in [above, below] {
+                                        fill_area.draw_styled(&background_style, target)?;
+                                    }
+                                }
+
+                                let decorations = carryover.decorations.into_iter().flatten();
+                                let decorations = decorations.map(|Styled { primitive, style }| {
+                                    primitive.intersection(&intersection).into_styled(style)
+                                });
+
+                                for decoration in decorations {
+                                    decoration.draw(target)?;
+                                }
+
+                                let column = intersection.y_extend(top, bottom);
+                                let above = column.above(&intersection);
+                                let below = column.below(&intersection);
+                                for clip_area in [above, below] {
+                                    let mut adapter = target.value_mapped(&colormap);
+                                    image.clipped(&clip_area).draw(&mut adapter)?;
+
+                                    let above = clip_area.above(&image_box);
+                                    let below = clip_area.below(&image_box);
+                                    for fill_area in [above, below] {
+                                        fill_area.draw_styled(&background_style, target)?;
+                                    }
+                                }
+
+                                image_box.right_of(&intersection)
+                            } else {
+                                image_box
+                            }
                         } else {
                             line_piece.draw_styled(&background_style, target)?;
 
@@ -250,7 +483,7 @@ macro_rules! impl_text_renderer {
                         previous_right = right.top_left.x.saturating_add_unsigned(right.size.width);
                     }
 
-                    if let Some(previous_image) = previous_image.take() {
+                    if let Some(previous_image) = previous_image.as_ref() {
                         let previous_image_box = previous_image.bounding_box();
                         let previous_right_half = previous_image_box.indent_to(previous_right);
                         let mut adapter = target.value_mapped(&colormap);
@@ -264,42 +497,43 @@ macro_rules! impl_text_renderer {
                         }
                     }
 
-                    let width = (x as i32).saturating_sub(right);
+                    let previous_image_colorable = previous_image.map(|previous_image| {
+                        PreviousImageColorable {
+                            previous_image,
+                            text_color: self.text_color(),
+                            background_color: self.background_color(),
+                        }
+                    });
+
+                    let next_position = Point::new(x as i32, position.y);
+                    let width = next_position.x.saturating_sub(right);
                     let width = width.try_into().unwrap_or_default();
                     let line_piece = Rectangle {
                         top_left: Point::new(right, top),
                         size: Size::new(width, height),
                     };
 
-                    let next_position = Point::new(x as i32, position.y);
                     line_piece.draw_styled(&background_style, target)?;
 
-                    let right = i32::max(x as i32, right);
+                    let width = right.saturating_sub(next_position.x);
+                    let width = width.try_into().unwrap_or_default();
+                    let line_piece = Rectangle {
+                        top_left: Point::new(next_position.x, top),
+                        size: Size::new(width, height),
+                    };
+
+                    let right = next_position.x.max(right);
                     let width = right.saturating_sub(position.x);
                     let width = width.try_into().unwrap_or_default();
+                    let decorations = self.decorations_drawn(position.x, y, width, target)?;
 
-                    if let Some(stroke_color) = self.underline_color() {
-                        let top = y.saturating_sub(self.font.underline.y_offset());
-                        let height = self.font.underline.stroke_width();
-                        let underline_style = PrimitiveStyle::with_fill(stroke_color);
-                        let underline = Rectangle {
-                            top_left: Point::new(position.x, top),
-                            size: Size::new(width, height),
+                    if self.carryover.borrow().is_none() {
+                        let carryover = Carryover {
+                            previous_image_colorable,
+                            decorations,
+                            line_piece
                         };
-
-                        underline.draw_styled(&underline_style, target)?;
-                    }
-
-                    if let Some(stroke_color) = self.strikethrough_color() {
-                        let top = y.saturating_sub(self.font.strikethrough.y_offset());
-                        let height = self.font.strikethrough.stroke_width();
-                        let strikethrough_style = PrimitiveStyle::with_fill(stroke_color);
-                        let strikethrough = Rectangle {
-                            top_left: Point::new(position.x, top),
-                            size: Size::new(width, height),
-                        };
-
-                        strikethrough.draw_styled(&strikethrough_style, target)?;
+                        self.carryover.replace(Some(carryover));
                     }
 
                     Ok(next_position)
@@ -326,31 +560,22 @@ macro_rules! impl_text_renderer {
                         size: Size::new(width, height),
                     };
 
-                    let next_position = Point::new(x as i32, position.y);
-                    line_piece.draw_styled(&background_style, target)?;
-
-                    if let Some(stroke_color) = self.underline_color() {
-                        let top = y.saturating_sub(self.font.underline.y_offset());
-                        let height = self.font.underline.stroke_width();
-                        let underline_style = PrimitiveStyle::with_fill(stroke_color);
-                        let underline = Rectangle {
-                            top_left: Point::new(position.x, top),
-                            size: Size::new(width, height),
-                        };
-
-                        underline.draw_styled(&underline_style, target)?;
+                    if let Some(carryover) = self.carryover.take() {
+                        carryover.redraw_whitespace(self, line_piece, target)?;
+                    } else {
+                        line_piece.draw_styled(&background_style, target)?;
                     }
 
-                    if let Some(stroke_color) = self.strikethrough_color() {
-                        let top = y.saturating_sub(self.font.strikethrough.y_offset());
-                        let height = self.font.strikethrough.stroke_width();
-                        let strikethrough_style = PrimitiveStyle::with_fill(stroke_color);
-                        let strikethrough = Rectangle {
-                            top_left: Point::new(position.x, top),
-                            size: Size::new(width, height),
-                        };
+                    let next_position = Point::new(x as i32, position.y);
+                    let decorations = self.decorations_drawn(position.x, y, width, target)?;
 
-                        strikethrough.draw_styled(&strikethrough_style, target)?;
+                    if self.carryover.borrow().is_none() {
+                        let carryover = Carryover {
+                            previous_image_colorable: None,
+                            decorations,
+                            line_piece
+                        };
+                        self.carryover.replace(Some(carryover));
                     }
 
                     Ok(next_position)